Added abstract item class
